+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # CHANGELOG
 
+## [1.32.1](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/compare/v1.32.0...v1.32.1) (2025-04-01)
+
+### 🐛 BUG FIXES
+
+* S3 이미지 업로드 로직 중복 제거 및 등록 로직 정리 ([45c4970](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/commit/45c49702271cff58d1918754c895a8c19b489368))
+
+## [1.32.0](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/compare/v1.31.0...v1.32.0) (2025-04-01)
+
+### 🚀 NEW FEATURES
+
+* Sentry update ([4370cd4](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/commit/4370cd4a2e3010e802f3ba70eeca3d34b57b20f8))
+
+## [1.31.0](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/compare/v1.30.0...v1.31.0) (2025-04-01)
+
+### 🚀 NEW FEATURES
+
+* Sentry user & S3 Image ([f0a5e65](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/commit/f0a5e65ec8f831a6c0431c55a72b8839a983c5f1))
+
+## [1.30.0](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/compare/v1.29.0...v1.30.0) (2025-04-01)
+
+### 🚀 NEW FEATURES
+
+* Sentry user & S3 Image ([2d064ee](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/commit/2d064ee3c8de6309d28ed8ffd946fe69c72be026))
+
 ## [1.29.0](https://github.com/100-hours-a-week/5-team-oishikunare-gangazido-fe/compare/v1.28.0...v1.29.0) (2025-04-01)
 
 ### 🚀 NEW FEATURES